+++ servo/utilities/__init__.py
@@ -4,3 +4,4 @@ from .pydantic import *
 from .yaml import *
 from .key_paths import *
 from .hashing import *
+from .subprocess import *